+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/GetUsersByRoleTask.Fields.Edit.cshtml
@@ -1,16 +1,16 @@
 @using OrchardCore.Roles.Workflows.ViewModels
-@model SelectUsersInRoleTaskViewModel
+@model GetUsersByRoleTaskViewModel
 
 <div class="mb-3" asp-validation-class-for="OutputKeyName">
     <label asp-for="OutputKeyName">@T["OutputKeyName"]</label>
     <input type="text" asp-for="OutputKeyName" class="form-control code" />
     <span asp-validation-for="OutputKeyName"></span>
-    <span class="hint">@T["The Output Key Name to save list. With Liquid support."]</span>
+    <span class="hint">@T["The designated key name for storing the users' list in the workflow output. This key enables later access to the list within the workflow.With Liquid support."]</span>
 </div>
 
 <div class="mb-3" asp-validation-class-for="Roles">
     <label asp-for="Roles">@T["Roles"]</label>
     @await Component.InvokeAsync("SelectRoles", new { selectedRoles = Model.Roles, htmlName = Html.NameFor(m => m.Roles), except = new[] { "Anonymous", "Authenticated" } })
     <span asp-validation-for="Roles"></span>
-    <span class="hint">@T["The Roles to search for. With Liquid support."]</span>
+    <span class="hint">@T["Choose the roles used to identify users during list generation."]</span>
 </div>
