+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/GetUsersByRoleTask.Fields.Thumbnail.cshtml
@@ -1,0 +1,2 @@
+<h4 class="card-title"><i class="fa-solid fa-user" aria-hidden="true"></i>@T["Get users in roles"]</h4>
+<p>@T["Get users in roles."]</p>
