--- src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/RemoveUserRoleTask.Fields.Edit.cshtml
@@ -1,16 +1,0 @@
-@using OrchardCore.Roles.Workflows.ViewModels
-@model RemoveUserRoleTaskViewModel
-
-<div class="mb-3" asp-validation-class-for="UserName">
-    <label asp-for="UserName">@T["UserName"]</label>
-    <input type="text" asp-for="UserName" class="form-control code" />
-    <span asp-validation-for="UserName"></span>
-    <span class="hint">@T["The User to update. With Liquid support."]</span>
-</div>
-
-<div class="mb-3" asp-validation-class-for="RoleName">
-    <label asp-for="RoleName">@T["RoleName"]</label>
-    <input type="text" asp-for="RoleName" class="form-control code" />
-    <span asp-validation-for="RoleName"></span>
-    <span class="hint">@T["The Role to remove. With Liquid support."]</span>
-</div>
