--- src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/RemoveUserRoleTask.Fields.Thumbnail.cshtml
@@ -1,2 +1,0 @@
-<h4 class="card-title"><i class="fa-solid fa-user" aria-hidden="true"></i>@T["Remove user from role"]</h4>
-<p>@T["Removes a user from a role."]</p>
