--- src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/SelectUsersInRoleTask.Fields.Thumbnail.cshtml
@@ -1,2 +1,0 @@
-<h4 class="card-title"><i class="fa-solid fa-user" aria-hidden="true"></i>@T["Select users in role"]</h4>
-<p>@T["Select users in role."]</p>
