+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/UnassignUserRoleTask.Fields.Edit.cshtml
@@ -1,0 +1,16 @@
+@using OrchardCore.Roles.Workflows.ViewModels
+@model UnassignUserRoleTaskViewModel
+
+<div class="mb-3" asp-validation-class-for="UserName">
+    <label asp-for="UserName">@T["UserName"]</label>
+    <input type="text" asp-for="UserName" class="form-control code" />
+    <span asp-validation-for="UserName"></span>
+    <span class="hint">@T["The User to update. With Liquid support."]</span>
+</div>
+
+<div class="mb-3" asp-validation-class-for="Roles">
+    <label asp-for="Roles">@T["Roles"]</label>
+    @await Component.InvokeAsync("SelectRoles", new { selectedRoles = Model.Roles, htmlName = Html.NameFor(m => m.Roles), except = new[] { "Anonymous", "Authenticated" } })
+    <span asp-validation-for="Roles"></span>
+    <span class="hint">@T["The Roles to remove."]</span>
+</div>
