+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/UnassignUserRoleTask.Fields.Thumbnail.cshtml
@@ -1,0 +1,2 @@
+<h4 class="card-title"><i class="fa-solid fa-user" aria-hidden="true"></i>@T["Unassign user from roles"]</h4>
+<p>@T["Unassign a user from roles."]</p>
