+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/Startup.cs
@@ -4,15 +4,14 @@ using OrchardCore.Roles.Workflows.Activities;
 using OrchardCore.Roles.Workflows.Drivers;
 using OrchardCore.Workflows.Helpers;
 
-namespace OrchardCore.Roles.Workflows
+namespace OrchardCore.Roles.Workflows;
+
+[RequireFeatures("OrchardCore.Workflows")]
+public class Startup : StartupBase
 {
-    [RequireFeatures("OrchardCore.Workflows")]
-    public class Startup : StartupBase
+    public override void ConfigureServices(IServiceCollection services)
     {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddActivity<RemoveUserRoleTask, RemoveUserRoleTaskDisplayDriver>();
-            services.AddActivity<SelectUsersInRoleTask, SelectUsersInRoleTaskDisplayDriver>();
-        }
+        services.AddActivity<UnassignUserRoleTask, UnassignUserRoleTaskDisplayDriver>();
+        services.AddActivity<GetUsersByRoleTask, GetUsersByRoleTaskDisplayDriver>();
     }
 }
