+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/ViewModels/GetUsersByRoleTaskViewModel.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+
+namespace OrchardCore.Roles.Workflows.ViewModels;
+
+public class GetUsersByRoleTaskViewModel
+{
+    [Required]
+    public string OutputKeyName { get; set; }
+
+    public IEnumerable<string> Roles { get; set; } = new List<string>();
+}
