--- src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/ViewModels/RemoveUserRoleTaskViewModel.cs
@@ -1,13 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace OrchardCore.Roles.Workflows.ViewModels
-{
-    public class RemoveUserRoleTaskViewModel
-    {
-        [Required]
-        public string UserName { get; set; }
-
-        [Required]
-        public string RoleName { get; set; }
-    }
-}
