--- src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/ViewModels/SelectUsersInRoleTaskViewModel.cs
@@ -1,13 +1,0 @@
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-
-namespace OrchardCore.Roles.Workflows.ViewModels
-{
-    public class SelectUsersInRoleTaskViewModel
-    {
-        [Required]
-        public string OutputKeyName { get; set; }
-
-        public IEnumerable<string> Roles { get; set; }
-    }
-}
