+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/ViewModels/UnassignUserRoleTaskViewModel.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+
+namespace OrchardCore.Roles.Workflows.ViewModels;
+
+public class UnassignUserRoleTaskViewModel
+{
+    [Required]
+    public string UserName { get; set; }
+
+    public IEnumerable<string> Roles { get; set; } = new List<string>();
+}
